feature: serving pdf file with orders details

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const {validationResult} = require('express-validator/check');
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product',{
@@ -25,8 +26,26 @@ exports.postAddProduct =  (req, res, next) => {
 
     const title = req.body.title;
     const price = req.body.price;
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const description = req.body.description;
+
+    if(!image) {
+        return res.status(422)
+        .render('admin/edit-product',{
+            pageTitle: 'Add product', 
+            path: '/admin/add-product', 
+            activeAddProduct: true, formsCSS: true, productCSS:true,
+            editting: false,
+            hasError: true,
+            product: {
+                title,
+                price,
+                description
+            },
+            errorMessage: 'Attached file is not an image',
+            validationErrors: []
+        }) 
+    }
 
     const errors = validationResult(req);
 
@@ -41,13 +60,14 @@ exports.postAddProduct =  (req, res, next) => {
             product: {
                 title,
                 price,
-                imageUrl,
                 description
             },
             errorMessage: errors.array()[0].msg,
             validationErrors: errors.array()
         }) 
     }
+
+    const imageUrl = image.path;
 
     const product = new Product({title, price, description, imageUrl, userId: user});
 
@@ -102,7 +122,7 @@ exports.postEditProduct = (req, res, next) => {
     const updatedPrice = req.body.price;
     const updatedTitle = req.body.title;
     const updatedDesc = req.body.description;
-    const updatedImage = req.body.imageUrl;
+    const image = req.file;
 
     const errors = validationResult(req);
 
@@ -118,7 +138,6 @@ exports.postEditProduct = (req, res, next) => {
                 _id: prodId,
                 title: updatedTitle,
                 price: updatedPrice,
-                imageUrl: updatedImage,
                 description: updatedDesc
             },
             errorMessage: errors.array()[0].msg,
@@ -134,7 +153,11 @@ exports.postEditProduct = (req, res, next) => {
         product.title = updatedTitle;
         product.price = updatedPrice;
         product.description = updatedDesc;
-        product.imageUrl = updatedImage;
+
+        if(image) {
+         fileHelper.deleteFile(product.imageUrl);
+         product.imageUrl = image.path;
+        }
 
         return product.save()    
         .then(() => {
@@ -170,7 +193,16 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.deleteOne({_id: prodId, userId: req.user._id})
+
+    Product.findById(prodId)
+    .then(product => {
+        if(!product) {
+            return next(new Error('Product not found'));
+        }
+        fileHelper.deleteFile(product.imageUrl);
+
+        return Product.deleteOne({_id: prodId, userId: req.user._id});
+    })
     .then(() => {
         res.redirect('/admin/products');
     })
